Changes from the reviewer
